+++ src/main/java/com/banking/card/repo/CardRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 public interface CardRepo extends CrudRepository<Card, Long> {
-    List<Card> findByCustomerId(int customerId);
+    List<Card> findByCustomerId(Long customerId);
 }
